% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{make_project}
\alias{make_project}
\title{Create DCL project template}
\usage{
make_project(path, ...)
}
\arguments{
\item{...}{}
}
\description{
Creates a project template with folders for raw data, scripts, data, EDA, and
reports. Creates a Makefile. Includes a script template.
}
